move NHM.MinerPluginToolkitV1.SgminerCommon code to TeamRedMiner
